i have changed the entire folder strucure and added 3 folders hooks components context and enhanced the navbar as said in issue
