+++ src/context/ThemeContext.js
@@ -1,0 +1,29 @@
+import React, { createContext, useContext, useCallback, useMemo } from 'react';
+import useLocalStorage from '../hooks/useLocalStorage';
+
+// ==================== Theme Context ====================
+const ThemeContext = createContext();
+
+export const ThemeProvider = ({ children }) => {
+  const [isDark, setIsDark] = useLocalStorage('darkMode', false);
+
+  const toggleTheme = useCallback(() => {
+    setIsDark(prev => !prev);
+  }, [setIsDark]);
+
+  const themeValue = useMemo(() => ({ isDark, toggleTheme }), [isDark, toggleTheme]);
+
+  return (
+    <ThemeContext.Provider value={themeValue}>
+      <div className={isDark ? 'dark' : ''}>
+        {children}
+      </div>
+    </ThemeContext.Provider>
+  );
+};
+
+export const useTheme = () => {
+  const context = useContext(ThemeContext);
+  if (!context) throw new Error('useTheme must be used within ThemeProvider');
+  return context;
+};
